+++ packages/meditrak-server/src/apiV2/export/exportSurveys/cellBuilders/ArithmeticConfigCellBuilder.js
@@ -3,19 +3,19 @@
  * Copyright (c) 2021 Beyond Essential Systems Pty Ltd
  */
 
- import { KeyValueCellBuilder } from './KeyValueCellBuilder';
+import { KeyValueCellBuilder } from './KeyValueCellBuilder';
 
+const SPECIAL_VALUE_PROCESSORS = [];
+export class ArithmeticConfigCellBuilder extends KeyValueCellBuilder {
+  // async processValue(value, key) {
+  //   return SPECIAL_VALUE_PROCESSORS[key]
+  //     ? SPECIAL_VALUE_PROCESSORS[key](value)
+  //     : this.fetchQuestionCode(value);
+  // }
 
- export class ArithmeticConfigCellBuilder extends KeyValueCellBuilder {
-  async processValue(value, key) {
-    return SPECIAL_VALUE_PROCESSORS[key]
-      ? SPECIAL_VALUE_PROCESSORS[key](value)
-      : this.fetchQuestionCode(value);
+  extractRelevantObject({ arithmetic }) {
+    return arithmetic;
   }
-
-   extractRelevantObject({ arithmetic }) {
-     return arithmetic;
-   }
 
   /**
    * (excel) answerDisplayText: SchFF00CalcTotal / SchFF00CalcTotalObtainable = $result
@@ -26,115 +26,115 @@
   async translateAnswerDisplayText(models, text, codes) {
     let translatedText = text;
     const questionCodeToId = await this.fetchQuestionCode()
-  
+
     for (const code of codes) {
       const questionId = questionCodeToId[code];
       translatedText = translatedText.replace(code, questionId);
     }
-  
+
     return translatedText;
-  };
- }
- 
- export const processArithmeticConfig = async (models, config) => {
-  const { formula, defaultValues, valueTranslation, answerDisplayText } = config;
-  const codes = getExpressionQuestionCodes(formula);
-
-  let translatedConfig = {
-    formula: await translateExpression(models, formula, codes),
-  };
-
-  if (defaultValues) {
-    translatedConfig = {
-      ...translatedConfig,
-      defaultValues: await translateDefaultValues(models, defaultValues),
-    };
   }
+}
 
-  if (valueTranslation) {
-    translatedConfig = {
-      ...translatedConfig,
-      valueTranslation: await translateValueTranslation(models, valueTranslation),
-    };
-  }
+// export const processArithmeticConfig = async (models, config) => {
+//   const { formula, defaultValues, valueTranslation, answerDisplayText } = config;
+//   const codes = getExpressionQuestionCodes(formula);
 
-  if (answerDisplayText) {
-    translatedConfig = {
-      ...translatedConfig,
-      answerDisplayText: await translateAnswerDisplayText(models, answerDisplayText, codes),
-    };
-  }
+//   let translatedConfig = {
+//     formula: await translateExpression(models, formula, codes),
+//   };
 
-  return translatedConfig;
-};
+//   if (defaultValues) {
+//     translatedConfig = {
+//       ...translatedConfig,
+//       defaultValues: await translateDefaultValues(models, defaultValues),
+//     };
+//   }
 
-/**
- * Excel config
- * defaultValues: SchFF00Section1Total:0,SchFF00Section1TotalObtainable:0
- *
- * will be converted to:
- * defaultValues:{
- *     5ed7558b61f76a6ba5003114:0,
- *     5ec5bf8961f76a18ad00d4f9:0
- * },
- * @param {*} models
- * @param {*} defaultValuesConfig
- */
-const translateDefaultValues = async (models, defaultValuesConfig) => {
-  const defaultValues = splitStringOnComma(defaultValuesConfig);
-  const translatedDefaultValues = {};
-  const codeToValue = {};
-  defaultValues.forEach(defaultValue => {
-    const [code, value] = splitStringOn(defaultValue, ':');
-    codeToValue[code] = value;
-  });
-  const questionCodeToId = await models.question.findIdByCode(Object.keys(codeToValue));
+//   if (valueTranslation) {
+//     translatedConfig = {
+//       ...translatedConfig,
+//       valueTranslation: await translateValueTranslation(models, valueTranslation),
+//     };
+//   }
 
-  Object.entries(codeToValue).forEach(([code, value]) => {
-    const questionId = questionCodeToId[code];
-    translatedDefaultValues[questionId] = value;
-  });
+//   if (answerDisplayText) {
+//     translatedConfig = {
+//       ...translatedConfig,
+//       answerDisplayText: await translateAnswerDisplayText(models, answerDisplayText, codes),
+//     };
+//   }
 
-  return translatedDefaultValues;
-};
+//   return translatedConfig;
+// };
 
-/**
- * Excel config
- * valueTranslation: SchCVD016a.Yes:3,SchCVD016a.No:0
- *
- * will be converted to:
- * valueTranslation: {
- *		5ed7558b61f76a6ba5003114: { Yes: 3, No: 0 }
- * }
- * @param {*} models
- * @param {*} valueTranslationConfig
- */
-const translateValueTranslation = async (models, valueTranslationConfig) => {
-  const valueTranslation = splitStringOnComma(valueTranslationConfig);
-  const translatedValueTranslation = {};
-  const codes = [
-    ...new Set( // Remove duplicated codes
-      valueTranslation.map(defaultValue => {
-        const [code] = splitStringOn(defaultValue, '.');
-        return code;
-      }),
-    ),
-  ];
-  const questionCodeToId = await models.question.findIdByCode(codes);
+// /**
+//  * Excel config
+//  * defaultValues: SchFF00Section1Total:0,SchFF00Section1TotalObtainable:0
+//  *
+//  * will be converted to:
+//  * defaultValues:{
+//  *     5ed7558b61f76a6ba5003114:0,
+//  *     5ec5bf8961f76a18ad00d4f9:0
+//  * },
+//  * @param {*} models
+//  * @param {*} defaultValuesConfig
+//  */
+// const translateDefaultValues = async (models, defaultValuesConfig) => {
+//   const defaultValues = splitStringOnComma(defaultValuesConfig);
+//   const translatedDefaultValues = {};
+//   const codeToValue = {};
+//   defaultValues.forEach(defaultValue => {
+//     const [code, value] = splitStringOn(defaultValue, ':');
+//     codeToValue[code] = value;
+//   });
+//   const questionCodeToId = await models.question.findIdByCode(Object.keys(codeToValue));
 
-  for (const translation of valueTranslation) {
-    const [code, value] = splitStringOn(translation, ':');
-    const [questionCode, questionOption] = splitStringOn(code, '.');
-    const questionId = questionCodeToId[questionCode];
+//   Object.entries(codeToValue).forEach(([code, value]) => {
+//     const questionId = questionCodeToId[code];
+//     translatedDefaultValues[questionId] = value;
+//   });
 
-    if (!translatedValueTranslation[questionId]) {
-      translatedValueTranslation[questionId] = {};
-    }
-    translatedValueTranslation[questionId][questionOption] = value;
-  }
+//   return translatedDefaultValues;
+// };
 
-  return translatedValueTranslation;
+// /**
+//  * Excel config
+//  * valueTranslation: SchCVD016a.Yes:3,SchCVD016a.No:0
+//  *
+//  * will be converted to:
+//  * valueTranslation: {
+//  *		5ed7558b61f76a6ba5003114: { Yes: 3, No: 0 }
+//  * }
+//  * @param {*} models
+//  * @param {*} valueTranslationConfig
+//  */
+// const translateValueTranslation = async (models, valueTranslationConfig) => {
+//   const valueTranslation = splitStringOnComma(valueTranslationConfig);
+//   const translatedValueTranslation = {};
+//   const codes = [
+//     ...new Set( // Remove duplicated codes
+//       valueTranslation.map(defaultValue => {
+//         const [code] = splitStringOn(defaultValue, '.');
+//         return code;
+//       }),
+//     ),
+//   ];
+//   const questionCodeToId = await models.question.findIdByCode(codes);
+
+//   for (const translation of valueTranslation) {
+//     const [code, value] = splitStringOn(translation, ':');
+//     const [questionCode, questionOption] = splitStringOn(code, '.');
+//     const questionId = questionCodeToId[questionCode];
+
+//     if (!translatedValueTranslation[questionId]) {
+//       translatedValueTranslation[questionId] = {};
+//     }
+//     translatedValueTranslation[questionId][questionOption] = value;
+//   }
+
+//   return translatedValueTranslation;
 
 
-};
+// };
