+++ packages/meditrak-server/src/tests/apiV2/export/exportSurveys/configCellBuilders/fixtures/ArithmeticConfigCellBuilderFixtures.js
@@ -1,0 +1,16 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+
+export const ARITHMETIC_TEST_CASES = [
+  {
+    description: 'works',
+    input: {
+      arithmetic: {
+        formula: '1',
+      }
+    },
+    expected: 'formula: 1',
+  },
+];
