+++ packages/meditrak-server/src/tests/apiV2/export/exportSurveys/configCellBuilders/fixtures/common.js
@@ -1,0 +1,11 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+
+export const QUESTIONS = [
+  {
+    code: 'hi',
+    id: 'test_1',
+  },
+];
