+++ packages/meditrak-server/src/tests/apiV2/export/exportSurveys/configCellBuilders/fixtures/index.js
@@ -1,0 +1,7 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
+ */
+
+export { ARITHMETIC_TEST_CASES } from './ArithmeticConfigCellBuilderFixtures';
+export { QUESIONS } from './common';
